validate-schema.js: Drop Gulp for raw Node.js
Gulp doesn't help us with much, and it makes passing arguments hard,
with complaints like:

  $ gulp validate src/Apache-2.0.xml
  [10:02:19] Using gulpfile ~/src/spdx/license-list-XML/gulpfile.js
  [10:02:19] Task 'src/Apache-2.0.xml' is not in your gulpfile
  ...

You can teach Gulp tasks about arguments [1], but the recommended
approach requires even more external dependencies.  Instead, this
commit just drops Gulp in favor raw Node.js.

It also adds the ability to validate individual files from the command
line:

  $ ./validate-schema.js src/Apache-2.0.xml
  validation complete 1 passed, 0 failed

which is useful for testing schema changes, since you can edit the XSD
and a single license while iterating on the schema and not be
distracted by test failures from all the other licenses.

[1]: https://gulpjs.org/recipes/pass-arguments-from-cli.html
